+++ test/test.py
@@ -1,5 +1,5 @@
 import unittest
-from game.tiles import Tile, BagTiles, Board, Player
+from game.tiles import Tile, BagTiles, Board, Player, Dictionary
 
 class TestTile(unittest.TestCase):
     def test_tile_creation(self):
@@ -107,6 +107,15 @@ class TestPlayer(unittest.TestCase):
         self.assertEqual(self.player.get_tile_value("A"), 1)
         self.assertEqual(self.player.get_tile_value("Z"), 10)
         self.assertEqual(self.player.get_tile_value("E"), 1)
+
+class TestDictionary(unittest.TestCase):
+    def test_dictionary(self):
+        dictionary = Dictionary('dictionaries/dictionary.txt')
+        self.assertTrue(dictionary.has_word('arbol'))
+
+    def test_word_false(self):
+        dictionary = Dictionary('dictionaries/dictionary.txt')
+        self.assertFalse(dictionary.has_word('willkommen'))
 
 if __name__ == '__main__':
     unittest.main()