+++ CHANGELOG.md
@@ -14,3 +14,7 @@ ADDED:
 ADDED:
 - Implementation of the Board and Player classes.
 - Added unit tests for the Board and Player classes.
+0.0.4 - 2023-08-23
+ADDED:
+- A dicitionary with a .txt file full of spanish word
+- A  Dictionary class  ans their respective test